重构，提取公有的类
