+++ core-impl/client/src/main/java/com/alipay/sofa/rpc/client/lb/WeightConsistentHashLoadBalancer.java
@@ -19,18 +19,11 @@ package com.alipay.sofa.rpc.client.lb;
 import com.alipay.sofa.rpc.bootstrap.ConsumerBootstrap;
 import com.alipay.sofa.rpc.client.AbstractLoadBalancer;
 import com.alipay.sofa.rpc.client.ProviderInfo;
-import com.alipay.sofa.rpc.common.utils.CommonUtils;
-import com.alipay.sofa.rpc.common.utils.StringUtils;
-import com.alipay.sofa.rpc.core.exception.SofaRpcRuntimeException;
+import com.alipay.sofa.rpc.client.Selector;
 import com.alipay.sofa.rpc.core.request.SofaRequest;
 import com.alipay.sofa.rpc.ext.Extension;
 
-import java.io.UnsupportedEncodingException;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.List;
-import java.util.Map;
-import java.util.TreeMap;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
@@ -69,153 +62,6 @@ public class WeightConsistentHashLoadBalancer extends AbstractLoadBalancer {
             selectorCache.put(key, selector);
         }
         return selector.select(request);
-    }
-
-    /**
-     * 选择器
-     */
-    private static class Selector {
-
-        /**
-         * The Hashcode.
-         */
-        private final int                         hashcode;
-
-        /**
-         * The Interface id.
-         */
-        private final String                      interfaceId;
-
-        /**
-         * The Method name.
-         */
-        private final String                      method;
-
-        /**
-         * 虚拟节点
-         */
-        private final TreeMap<Long, ProviderInfo> virtualNodes;
-
-        /**
-         * Instantiates a new Selector .
-         *
-         * @param interfaceId the interface id
-         * @param method      the method
-         * @param actualNodes the actual nodes
-         */
-        public Selector(String interfaceId, String method, List<ProviderInfo> actualNodes) {
-            this(interfaceId, method, actualNodes, actualNodes.hashCode());
-        }
-
-        /**
-         * Instantiates a new Selector.
-         *
-         * @param interfaceId the interface id
-         * @param method      the method
-         * @param actualNodes the actual nodes
-         * @param hashcode    the hashcode
-         */
-        public Selector(String interfaceId, String method, List<ProviderInfo> actualNodes, int hashcode) {
-            this.interfaceId = interfaceId;
-            this.method = method;
-            this.hashcode = hashcode;
-            // 创建虚拟节点环 （provider创建虚拟节点数 =  真实节点权重 * 32）
-            this.virtualNodes = new TreeMap<Long, ProviderInfo>();
-            // 设置越大越慢，精度越高
-            int num = 32;
-            for (ProviderInfo providerInfo : actualNodes) {
-                for (int i = 0; i < num * providerInfo.getWeight() / 4; i++) {
-                    byte[] digest = messageDigest(providerInfo.getHost() + providerInfo.getPort() + i);
-                    for (int h = 0; h < 4; h++) {
-                        long m = hash(digest, h);
-                        virtualNodes.put(m, providerInfo);
-                    }
-                }
-            }
-        }
-
-        /**
-         * Select provider.
-         *
-         * @param request the request
-         * @return the provider
-         */
-        public ProviderInfo select(SofaRequest request) {
-            String key = buildKeyOfHash(request.getMethodArgs());
-            byte[] digest = messageDigest(key);
-            return selectForKey(hash(digest, 0));
-        }
-
-        /**
-         * 获取第一参数作为hash的key
-         *
-         * @param args the args
-         * @return the string
-         */
-        private String buildKeyOfHash(Object[] args) {
-            if (CommonUtils.isEmpty(args)) {
-                return StringUtils.EMPTY;
-            } else {
-                return StringUtils.toString(args[0]);
-            }
-        }
-
-        /**
-         * Select for key.
-         *
-         * @param hash the hash
-         * @return the provider
-         */
-        private ProviderInfo selectForKey(long hash) {
-            Map.Entry<Long, ProviderInfo> entry = virtualNodes.ceilingEntry(hash);
-            if (entry == null) {
-                entry = virtualNodes.firstEntry();
-            }
-            return entry.getValue();
-        }
-
-        /**
-         * 换算法？ MD5  SHA-1 MurMurHash???
-         *
-         * @param value the value
-         * @return the byte []
-         */
-        private byte[] messageDigest(String value) {
-            MessageDigest md5;
-            try {
-                md5 = MessageDigest.getInstance("MD5");
-                md5.update(value.getBytes("UTF-8"));
-                return md5.digest();
-            } catch (NoSuchAlgorithmException e) {
-                throw new SofaRpcRuntimeException("No such algorithm named md5", e);
-            } catch (UnsupportedEncodingException e) {
-                throw new SofaRpcRuntimeException("Unsupported encoding of" + value, e);
-            }
-        }
-
-        /**
-         * Hash long.
-         *
-         * @param digest the digest
-         * @param index  the number
-         * @return the long
-         */
-        private long hash(byte[] digest, int index) {
-            long f = ((long) (digest[3 + index * 4] & 0xFF) << 24)
-                | ((long) (digest[2 + index * 4] & 0xFF) << 16)
-                | ((long) (digest[1 + index * 4] & 0xFF) << 8)
-                | (digest[index * 4] & 0xFF);
-            return f & 0xFFFFFFFFL;
-        }
-
-        /**
-         * Gets hash code.
-         *
-         * @return the hash code
-         */
-        public int getHashCode() {
-            return hashcode;
-        }
     }
 
 }
